project picker improvements
